+++ api/models/reservations.js
@@ -972,6 +972,7 @@ const addRoomToReservation = async (requestId, reservationId, numberOfPeople, ro
   try {
     await client.query('BEGIN');
 
+
     // Update the number_of_people in the reservations table
     const updateReservationQuery = `
       UPDATE reservations
@@ -979,12 +980,12 @@ const addRoomToReservation = async (requestId, reservationId, numberOfPeople, ro
           updated_by = $2
       WHERE id = $3::UUID;
     `;
-    await pool.query(updateReservationQuery, [numberOfPeople, userId, reservationId]);
+    await client.query(updateReservationQuery, [numberOfPeople, userId, reservationId]);
 
     // Copy one existing room_id in the reservation_details table
     const copyRoomQuery = `
-      INSERT INTO reservation_details (hotel_id, reservation_id, date, room_id, plans_global_id, plans_hotel_id, number_of_people, price, created_by, updated_by)
-      SELECT hotel_id, reservation_id, date, $1, NULL, NULL, $2, 0, created_by, $3
+      INSERT INTO reservation_details (hotel_id, reservation_id, date, room_id, plans_global_id, plans_hotel_id, number_of_people, price, created_by, updated_by, billable)
+      SELECT hotel_id, reservation_id, date, $1, NULL, NULL, $2, 0, created_by, $3, billable
       FROM reservation_details
       WHERE (hotel_id, reservation_id, room_id) IN (
         SELECT hotel_id, reservation_id, room_id
@@ -994,7 +995,7 @@ const addRoomToReservation = async (requestId, reservationId, numberOfPeople, ro
       )
       RETURNING *;
     `;
-    const result = await pool.query(copyRoomQuery, [roomId, numberOfPeople, userId, reservationId]);
+    const result = await client.query(copyRoomQuery, [roomId, numberOfPeople, userId, reservationId]);
 
     await client.query('COMMIT');
     return result.rows[0];